chapter07/dictionary.go
package dictionary

var (
	ErrKeyNotFound  = DictionaryErr("Key not found.")
	ErrKeyExists    = DictionaryErr("Key exists.")
	ErrKeyNotExists = DictionaryErr("Key not exists.")
)

type DictionaryErr string

func (e DictionaryErr) Error() string {
	return string(e)
}

type Dictionary map[string]string

func (d Dictionary) Search(key string) (string, error) {
	value, ok := d[key]
	if !ok {
		return "", ErrKeyNotFound
	}
	return value, nil
}

func (d Dictionary) Add(key, value string) error {
	_, err := d.Search(key)

	switch err {
	case ErrKeyNotFound:
		d[key] = value
	case nil:
		return ErrKeyExists
	default:
		return err
	}

	return nil
}

func (d Dictionary) Update(key, value string) error {
	_, err := d.Search(key)

	switch err {
	case ErrKeyNotFound:
		return ErrKeyNotExists
	case nil:
		d[key] = value
	default:
		return err
	}

	return nil
}

func (d Dictionary) Delete(key string) error {
	_, err := d.Search(key)

	switch err {
	case ErrKeyNotFound:
		return ErrKeyNotExists
	case nil:
		delete(d, key)
	default:
		return err
	}

	return nil

}
